requirements.txt
openai
dotenv
httpx
retell-sdk
uvicorn
fastapi
fastapi[standard]
'crewai[tools]'